backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
